Buttons created.
Rather than single prepare_cards_btn now have assign_cards_btns.  This is a list of lists of buttons (one list of buttons for each player).
Buttons created, but not yet active.
Also created method to change text within a button.  Initially text is set # or run #, but can change it to the contents of that set (if we choose to).