visible_cards needed by handView.
Ran into 2 related problems, both traceable to the fact that Liverpool allows you to play on others sets and groups.
1. played_cards in HandAndFoot can be totally local, but it cannot be local in Liverpool.
2. I had been using different indexing scheme in preparing cards then what the server uses for visible cards. Need to fix that.
These problems need to be fixed in order to do rule checking when adding cards to others' runs or sets.
--
Changes in this commit pass more variables from RunClient to handView.  But problems still need to be debugged.