Oops...
Realized that initial structure I chose of having assign buttons keyed on tuple was causing needless complications. Those buttons should follow same structure as visible_cards, and be a list of dictionaries.  This makes a lot of the code in this branch extraneous.. Save it and go back to Liverpool branch and work from there.