visible_cards_now is longest vcards
Realized best way to determine which players v_cards should be broadcast to show what cards have been played was to use the one with the most cards played.
This removes need to consolidate all the cards in each players hand.
this commit still has a few bugs,but much closer...