verified try: statement needed
note for reviewer: host name more stable than port#, so changed code so ask for host and port # separately. This removes the need for player to type in host name when it's default.

Also verified the if/while statements used to set player_index are needed
(loop goes through a couple of times)